+++ client/src/api/groups.js
@@ -3,3 +3,7 @@ import { remove } from './base'
 export function deleteGroup (id) {
   return remove(`/groups/${id}`)
 }
+
+export function removeMember (groupId, memberId) {
+  return remove(`/groups/${groupId}/members/${memberId}`)
+}