+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Allow admins of the working group "Redaktion" to edit blog posts #34 !2061 @alex.simm
 
 - "New event" page requires login !2058 @Buntelrus
+- Add a button to the group member list for removing members !2073 @alex.simm @chriswalg
+
 ## Bugfixes
 - Fixed the list of responsible store members in REST responses !2033 #1124 @alex.simm
 - Fixed for content edit !2062 @fs_k